checkpoint: renames "Attempt/adapter" to "Attempt/Adapted"
