+++ src/library/Attempt/Adapted/Config.ts
