+++ src/library/Attempt/Adapted/asynchronous.ts
