+++ src/library/Attempt/Adapted/index.ts
