+++ src/library/Attempt/Adapted/synchronous.ts
